+++ sentry-android-core/src/main/java/io/sentry/android/core/AnrV2Integration.java
@@ -22,6 +22,7 @@ import io.sentry.android.core.internal.util.NativeEventUtils;
 import io.sentry.hints.AbnormalExit;
 import io.sentry.hints.Backfillable;
 import io.sentry.hints.BlockingFlushHint;
+import io.sentry.protocol.ArtContext;
 import io.sentry.protocol.DebugImage;
 import io.sentry.protocol.DebugMeta;
 import io.sentry.protocol.Message;
@@ -173,6 +174,9 @@ public class AnrV2Integration implements Integration, Closeable {
           debugMeta.setImages(result.debugImages);
           event.setDebugMeta(debugMeta);
         }
+        if (result.artContext != null) {
+          event.getContexts().setArt(result.artContext);
+        }
       }
       event.setLevel(SentryLevel.FATAL);
       event.setTimestamp(DateUtils.getDateTime(anrTimestamp));
@@ -209,6 +213,7 @@ public class AnrV2Integration implements Integration, Closeable {
 
         final @NotNull List<SentryThread> threads = threadDumpParser.getThreads();
         final @NotNull List<DebugImage> debugImages = threadDumpParser.getDebugImages();
+        final @Nullable ArtContext artContext = threadDumpParser.getArtContext();
 
         if (threads.isEmpty()) {
           // if the list is empty this means the system failed to capture a proper thread dump of
@@ -217,7 +222,7 @@ public class AnrV2Integration implements Integration, Closeable {
           // fall back to not reporting them
           return new ParseResult(ParseResult.Type.NO_DUMP);
         }
-        return new ParseResult(ParseResult.Type.DUMP, dump, threads, debugImages);
+        return new ParseResult(ParseResult.Type.DUMP, dump, threads, debugImages, artContext);
       } catch (Throwable e) {
         options.getLogger().log(SentryLevel.WARNING, "Failed to parse ANR thread dump", e);
         return new ParseResult(ParseResult.Type.ERROR, dump);
@@ -286,15 +291,17 @@ public class AnrV2Integration implements Integration, Closeable {
     }
 
     final Type type;
-    final byte[] dump;
+    final @Nullable byte[] dump;
     final @Nullable List<SentryThread> threads;
     final @Nullable List<DebugImage> debugImages;
+    final @Nullable ArtContext artContext;
 
     ParseResult(final @NotNull Type type) {
       this.type = type;
       this.dump = null;
       this.threads = null;
       this.debugImages = null;
+      this.artContext = null;
     }
 
     ParseResult(final @NotNull Type type, final byte[] dump) {
@@ -302,17 +309,20 @@ public class AnrV2Integration implements Integration, Closeable {
       this.dump = dump;
       this.threads = null;
       this.debugImages = null;
+      this.artContext = null;
     }
 
     ParseResult(
         final @NotNull Type type,
         final byte[] dump,
         final @Nullable List<SentryThread> threads,
-        final @Nullable List<DebugImage> debugImages) {
+        final @Nullable List<DebugImage> debugImages,
+        final @Nullable ArtContext artContext) {
       this.type = type;
       this.dump = dump;
       this.threads = threads;
       this.debugImages = debugImages;
+      this.artContext = artContext;
     }
   }
 }