+++ sentry/src/main/java/io/sentry/protocol/Contexts.java
@@ -64,6 +64,8 @@ public class Contexts implements JsonSerializable {
           this.setResponse(new Response((Response) value));
         } else if (Spring.TYPE.equals(entry.getKey()) && value instanceof Spring) {
           this.setSpring(new Spring((Spring) value));
+        } else if (ArtContext.TYPE.equals(entry.getKey()) && value instanceof ArtContext) {
+          this.setArt(new ArtContext((ArtContext) value));
         } else {
           this.put(entry.getKey(), value);
         }
@@ -179,6 +181,14 @@ public class Contexts implements JsonSerializable {
 
   public void setSpring(final @NotNull Spring spring) {
     this.put(Spring.TYPE, spring);
+  }
+
+  public @Nullable ArtContext getArt() {
+    return toContextType(ArtContext.TYPE, ArtContext.class);
+  }
+
+  public void setArt(final @NotNull ArtContext art) {
+    this.put(ArtContext.TYPE, art);
   }
 
   public @Nullable FeatureFlags getFeatureFlags() {
@@ -346,6 +356,9 @@ public class Contexts implements JsonSerializable {
             break;
           case Spring.TYPE:
             contexts.setSpring(new Spring.Deserializer().deserialize(reader, logger));
+            break;
+          case ArtContext.TYPE:
+            contexts.setArt(new ArtContext.Deserializer().deserialize(reader, logger));
             break;
           case FeatureFlags.TYPE:
             contexts.setFeatureFlags(new FeatureFlags.Deserializer().deserialize(reader, logger));