+++ sentry/src/test/java/io/sentry/protocol/ContextsSerializationTest.kt
@@ -25,6 +25,7 @@ class ContextsSerializationTest {
         setResponse(ResponseSerializationTest.Fixture().getSut())
         setTrace(SpanContextSerializationTest.Fixture().getSut())
         setSpring(SpringSerializationTest.Fixture().getSut())
+        setArt(ArtContextSerializationTest.Fixture().getSut())
         setFeatureFlags(FeatureFlagsSerializationTest.Fixture().getSut())
       }
   }